doc/requirements.txt
sphinx_issues
cython
setuptools_scm
sphinx_better_subsection
